requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
tensorflow
joblib
comet-ml
dvc